+++ frontend/src/components/Chat/sub-components/ChatInputTypingIndicator/index.tsx
